Do not reset scale to 1 if value is forcefed
Velocity [resets scale](https://github.com/julianshapiro/velocity/blob/master/jquery.velocity.js#L874-L879) on android to 1 regardless of a forcefed value.

I forcefeed the value that I intend the animation to start from, but instead there is a flash of the element jumping to scale 1. If there is a forcefed value, velocity should accept it instead of overriding it.