Fix animateParentHeight height calc when box-border
animateParentHeight was over-calculating the height of the children elements
which have the box-sizing set to box border. This resulted in the parent
being animated too tall: http://jsbin.com/xakipigino/1/edit

For elements with box-sizing: "border-box", the height value returned by
Velocity.CSS.getPropertyValue(element, 'height') already includes
padding and border. So in these cases, padding and border should no be
added when calculating the total height occupied the element.