Improves jQuery-wrapped elements check in isWrapped.
Checking for the existence of 'jquery' on an object is not as reliable
as looking at the constructor's prototype of the object. Doing that
affords us the opportunity to use 'jquery' for other means—and not to
determine that it's a jQuery-wrapped object. Of course, this merely
moves the problem up the chain. Regardless, it's a little more cemented
and will work for an even larger majority of cases.