WIP BUILD - TESTS NOT WORKING Hopefully the caps made it obvious. The internals have changed significantly, so most of the tests can no longer work as they're checking internals rather than externals. This is completely clear from jQuery, and uses more ES6/7 features that TypeScript cleans up for us. There are several major slowdowns simply because it's mid-change (but I wanted to get something out to show where things are going). I've tried to mark some data structures as @deprecated where I remember to (things like `display` and `visibility` are going to become plain properties - they might stay in the options, but only as an alias).
Performance is currently pretty bad - not least because things like display and visibility are being set on every frame (the biggest slowdown is currently the css setting call as it's being overused multiple times on every frame on every animation).

Internal caching is currently disabled - not quite sure what's going wrong on it, but the only downside is a bit of memory waste.

Internal changes include the `tick`, `progress` and `complete` being on different loops (so doing stuff in them won't slow down animations), and "new" animations are now appended to the list and take effect from the next tick - anything "new" in the current tick (including the fake one that's called when adding anything) will get it's properties expanded then (convert from the loose combinations allowed, to the specific format needed internally).

Loops and repeats (yes, there's a new `repeat:10` option - same as loop but without reversing) are now lazy - it doesn't add a pair copies of the `"repeat"` call on the end of the animation queue for every number (so it's now possible to `loop:Number.POSITIVE_INFINITY` etc).

I also have a plan / design for overloading API for the `properties` - to start with it will get `scroll`, `display` and `visibility` added, the latter two to behave the same way as custom code within the tick currently behaves - hide at the end, show at the beginning. Scroll will get `scrollTop` and `scrollLeft` added (scrolling the window will automatically use the correct method), so you can scroll a specific item - I'd like my eventual plan to allow it to scroll to a specific item by passing that as the property value, but for now it'll take numbers. Other current types will be added the same way (using a lot of the currently commented out code from the old way I expect).

Just a quick note that this is still most definitely BETA - so don't use in production etc ;-)