Add "window" to access to the 'navigator' property (#803)
* Add "window" to access to the 'navigator' property

I don't kow the reason why but in my case I'm getting an error "navigator is not defined".
Adding "window" in front of it, resolves my problem.
As adding or not "window" should not affect the code, I assume that putting it should not be an issue for anyone but it will be appreciated by me ;)

* update min and ui.min files